extract search impl into separate files
