+++ src/RA_Integration.vcxproj
@@ -100,6 +100,8 @@
     <ClCompile Include="services\Initialization.cpp" />
     <ClCompile Include="services\PerformanceCounter.cpp" />
     <ClCompile Include="services\SearchResults.cpp" />
+    <ClCompile Include="services\Search\MemBlock.cpp" />
+    <ClCompile Include="services\search\SearchImpl.cpp" />
     <ClCompile Include="ui\drawing\gdi\GDIBitmapSurface.cpp" />
     <ClCompile Include="ui\drawing\gdi\GDISurface.cpp" />
     <ClCompile Include="ui\drawing\gdi\ImageRepository.cpp" />
@@ -266,6 +268,27 @@
     <ClInclude Include="services\Initialization.hh" />
     <ClInclude Include="services\IThreadPool.hh" />
     <ClInclude Include="services\PerformanceCounter.hh" />
+    <ClInclude Include="services\Search\MemBlock.hh" />
+    <ClInclude Include="services\search\SearchImpl.hh" />
+    <ClInclude Include="services\search\SearchImpl_16bit.hh" />
+    <ClInclude Include="services\search\SearchImpl_16bit_aligned.hh" />
+    <ClInclude Include="services\search\SearchImpl_16bit_be.hh" />
+    <ClInclude Include="services\search\SearchImpl_16bit_be_aligned.hh" />
+    <ClInclude Include="services\search\SearchImpl_24bit.hh" />
+    <ClInclude Include="services\search\SearchImpl_32bit.hh" />
+    <ClInclude Include="services\search\SearchImpl_32bit_aligned.hh" />
+    <ClInclude Include="services\search\SearchImpl_32bit_be.hh" />
+    <ClInclude Include="services\search\SearchImpl_32bit_be_aligned.hh" />
+    <ClInclude Include="services\search\SearchImpl_4bit.hh" />
+    <ClInclude Include="services\search\SearchImpl_8bit.hh" />
+    <ClInclude Include="services\search\SearchImpl_asciitext.hh" />
+    <ClInclude Include="services\search\SearchImpl_bitcount.hh" />
+    <ClInclude Include="services\search\SearchImpl_double32.hh" />
+    <ClInclude Include="services\search\SearchImpl_double32_be.hh" />
+    <ClInclude Include="services\search\SearchImpl_float.hh" />
+    <ClInclude Include="services\search\SearchImpl_float_be.hh" />
+    <ClInclude Include="services\search\SearchImpl_mbf32.hh" />
+    <ClInclude Include="services\search\SearchImpl_mbf32_le.hh" />
     <ClInclude Include="services\ServiceLocator.hh" />
     <ClInclude Include="services\SearchResults.h" />
     <ClInclude Include="services\TextReader.hh" />
