+++ src/RA_Integration.vcxproj.filters
@@ -52,6 +52,9 @@
     <Filter Include="Data\Models">
       <UniqueIdentifier>{c9c038f9-c4bb-436c-8416-12f91df44904}</UniqueIdentifier>
     </Filter>
+    <Filter Include="Services\Search">
+      <UniqueIdentifier>{e978c08c-fedb-4ead-af63-0420b26291bb}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="RA_md5factory.cpp">
@@ -419,6 +422,12 @@
     </ClCompile>
     <ClCompile Include="services\AchievementLogicSerializer.cpp">
       <Filter>Services</Filter>
+    </ClCompile>
+    <ClCompile Include="services\Search\MemBlock.cpp">
+      <Filter>Services\Search</Filter>
+    </ClCompile>
+    <ClCompile Include="services\search\SearchImpl.cpp">
+      <Filter>Services\Search</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -1000,6 +1009,69 @@
     </ClInclude>
     <ClInclude Include="services\AchievementLogicSerializer.hh">
       <Filter>Services</Filter>
+    </ClInclude>
+    <ClInclude Include="services\Search\MemBlock.hh">
+      <Filter>Services\Search</Filter>
+    </ClInclude>
+    <ClInclude Include="services\search\SearchImpl.hh">
+      <Filter>Services\Search</Filter>
+    </ClInclude>
+    <ClInclude Include="services\search\SearchImpl_4bit.hh">
+      <Filter>Services\Search</Filter>
+    </ClInclude>
+    <ClInclude Include="services\search\SearchImpl_8bit.hh">
+      <Filter>Services\Search</Filter>
+    </ClInclude>
+    <ClInclude Include="services\search\SearchImpl_16bit.hh">
+      <Filter>Services\Search</Filter>
+    </ClInclude>
+    <ClInclude Include="services\search\SearchImpl_16bit_aligned.hh">
+      <Filter>Services\Search</Filter>
+    </ClInclude>
+    <ClInclude Include="services\search\SearchImpl_16bit_be.hh">
+      <Filter>Services\Search</Filter>
+    </ClInclude>
+    <ClInclude Include="services\search\SearchImpl_16bit_be_aligned.hh">
+      <Filter>Services\Search</Filter>
+    </ClInclude>
+    <ClInclude Include="services\search\SearchImpl_24bit.hh">
+      <Filter>Services\Search</Filter>
+    </ClInclude>
+    <ClInclude Include="services\search\SearchImpl_32bit.hh">
+      <Filter>Services\Search</Filter>
+    </ClInclude>
+    <ClInclude Include="services\search\SearchImpl_32bit_aligned.hh">
+      <Filter>Services\Search</Filter>
+    </ClInclude>
+    <ClInclude Include="services\search\SearchImpl_32bit_be.hh">
+      <Filter>Services\Search</Filter>
+    </ClInclude>
+    <ClInclude Include="services\search\SearchImpl_32bit_be_aligned.hh">
+      <Filter>Services\Search</Filter>
+    </ClInclude>
+    <ClInclude Include="services\search\SearchImpl_asciitext.hh">
+      <Filter>Services\Search</Filter>
+    </ClInclude>
+    <ClInclude Include="services\search\SearchImpl_bitcount.hh">
+      <Filter>Services\Search</Filter>
+    </ClInclude>
+    <ClInclude Include="services\search\SearchImpl_double32.hh">
+      <Filter>Services\Search</Filter>
+    </ClInclude>
+    <ClInclude Include="services\search\SearchImpl_double32_be.hh">
+      <Filter>Services\Search</Filter>
+    </ClInclude>
+    <ClInclude Include="services\search\SearchImpl_float.hh">
+      <Filter>Services\Search</Filter>
+    </ClInclude>
+    <ClInclude Include="services\search\SearchImpl_float_be.hh">
+      <Filter>Services\Search</Filter>
+    </ClInclude>
+    <ClInclude Include="services\search\SearchImpl_mbf32.hh">
+      <Filter>Services\Search</Filter>
+    </ClInclude>
+    <ClInclude Include="services\search\SearchImpl_mbf32_le.hh">
+      <Filter>Services\Search</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
