+++ src/services/SearchResults.h
@@ -6,6 +6,8 @@
 
 #include "data\context\EmulatorContext.hh"
 
+#include "search\MemBlock.hh"
+
 namespace ra {
 namespace services {
 
@@ -42,153 +44,14 @@ enum class SearchFilterType
     InitialValue,
 };
 
-namespace impl {
+namespace search { class SearchImpl; }
 
-class MemBlock
+struct SearchResult
 {
-public:
-    explicit MemBlock(_In_ unsigned int nAddress, _In_ unsigned int nSize, _In_ unsigned int nMaxAddresses) noexcept :
-        m_nFirstAddress(nAddress),
-        m_nBytesSize(nSize),
-        m_nMatchingAddresses(nMaxAddresses),
-        m_nMaxAddresses(nMaxAddresses)
-    {
-        if (nSize > sizeof(m_vBytes))
-            m_pBytes = new (std::nothrow) uint8_t[nSize];
-    }
-
-    MemBlock(const MemBlock& other) noexcept : MemBlock(other.m_nFirstAddress, other.m_nBytesSize, other.m_nMaxAddresses)
-    {
-        if (m_nBytesSize > sizeof(m_vBytes))
-            std::memcpy(m_pBytes, other.m_pBytes, m_nBytesSize);
-        else
-            std::memcpy(m_vBytes, other.m_vBytes, sizeof(m_vBytes));
-
-        m_nMatchingAddresses = other.m_nMatchingAddresses;
-        if (!AreAllAddressesMatching())
-        {
-            if ((m_nMatchingAddresses + 7) / 8 > sizeof(m_vAddresses))
-            {
-                auto* pAddresses = AllocateMatchingAddresses();
-                if (pAddresses != nullptr)
-                    std::memcpy(pAddresses, other.m_pAddresses, (m_nMatchingAddresses + 7) / 8);
-            }
-            else
-            {
-                std::memcpy(m_vAddresses, other.m_vAddresses, sizeof(m_vAddresses));
-            }
-        }
-    }
-
-    MemBlock& operator=(const MemBlock&) noexcept = delete;
-
-    MemBlock(MemBlock&& other) noexcept :
-        m_nFirstAddress(other.m_nFirstAddress),
-        m_nBytesSize(other.m_nBytesSize),
-        m_nMatchingAddresses(other.m_nMatchingAddresses),
-        m_nMaxAddresses(other.m_nMaxAddresses)
-    {
-        if (other.m_nBytesSize > sizeof(m_vBytes))
-        {
-            m_pBytes = other.m_pBytes;
-            other.m_pBytes = nullptr;
-            other.m_nBytesSize = 0;
-        }
-        else
-        {
-            std::memcpy(m_vBytes, other.m_vBytes, sizeof(m_vBytes));
-        }
-
-        if (!AreAllAddressesMatching())
-        {
-            if ((m_nMatchingAddresses + 7) / 8 > sizeof(m_vAddresses))
-            {
-                m_pAddresses = other.m_pAddresses;
-                other.m_pAddresses = nullptr;
-                other.m_nMatchingAddresses = 0;
-            }
-            else
-            {
-                std::memcpy(m_vAddresses, other.m_vAddresses, sizeof(m_vAddresses));
-            }
-        }
-    }
-
-    MemBlock& operator=(MemBlock&&) noexcept = delete;
-
-    ~MemBlock() noexcept
-    {
-        if (m_nBytesSize > sizeof(m_vBytes))
-            delete[] m_pBytes;
-        if (!AreAllAddressesMatching() && (m_nMatchingAddresses + 7) / 8 > sizeof(m_vAddresses))
-            delete[] m_pAddresses;
-    }
-
-    uint8_t* GetBytes() noexcept { return (m_nBytesSize > sizeof(m_vBytes)) ? m_pBytes : &m_vBytes[0]; }
-    const uint8_t* GetBytes() const noexcept { return (m_nBytesSize > sizeof(m_vBytes)) ? m_pBytes : &m_vBytes[0]; }
-
-    ra::ByteAddress GetFirstAddress() const noexcept { return m_nFirstAddress; }
-    unsigned int GetBytesSize() const noexcept { return m_nBytesSize; }
-    unsigned int GetMaxAddresses() const noexcept { return m_nMaxAddresses; }
-
-    bool ContainsAddress(ra::ByteAddress nAddress) const noexcept;
-
-    void SetMatchingAddresses(std::vector<ra::ByteAddress>& vAddresses, gsl::index nFirstIndex, gsl::index nLastIndex);
-    void CopyMatchingAddresses(const MemBlock& pSource);
-    void ExcludeMatchingAddress(ra::ByteAddress nAddress);
-    bool ContainsMatchingAddress(ra::ByteAddress nAddress) const;
-
-    unsigned int GetMatchingAddressCount() const noexcept { return m_nMatchingAddresses; }
-    ra::ByteAddress GetMatchingAddress(gsl::index nIndex) const noexcept;
-    bool AreAllAddressesMatching() const noexcept { return m_nMatchingAddresses == m_nMaxAddresses; }
-
-    const uint8_t* GetMatchingAddressPointer() const noexcept
-    {
-        if (AreAllAddressesMatching())
-            return nullptr;
-
-        const auto nAddressesSize = (m_nMaxAddresses + 7) / 8;
-        return (nAddressesSize > sizeof(m_vAddresses)) ? m_pAddresses : &m_vAddresses[0];
-    }
-
-    bool HasMatchingAddress(const uint8_t* pMatchingAddresses, ra::ByteAddress nAddress) const noexcept
-    {
-        if (!pMatchingAddresses)
-            return true;
-
-        const auto nIndex = nAddress - m_nFirstAddress;
-        if (nIndex >= m_nMaxAddresses)
-            return false;
-
-        const auto nBit = 1 << (nIndex & 7);
-        return (pMatchingAddresses[nIndex >> 3] & nBit);
-    }
-
-private:
-    uint8_t* AllocateMatchingAddresses() noexcept;
-
-    union // 8 bytes
-    {
-        uint8_t m_vBytes[8]{};
-        uint8_t* m_pBytes;
-    };
-
-    union // 8 bytes
-    {
-        uint8_t m_vAddresses[8]{};
-        uint8_t* m_pAddresses;
-    };
-
-    unsigned int m_nBytesSize;         // 4 bytes
-    ra::ByteAddress m_nFirstAddress;   // 4 bytes
-    unsigned int m_nMatchingAddresses; // 4 bytes
-    unsigned int m_nMaxAddresses;      // 4 bytes
+    ra::ByteAddress nAddress{};
+    unsigned int nValue{};
+    MemSize nSize{};
 };
-static_assert(sizeof(MemBlock) <= 32, "sizeof(MemBlock) is incorrect");
-
-class SearchImpl;
-
-} // namespace impl
 
 class SearchResults
 {
@@ -241,26 +104,19 @@ public:
     /// </summary>
     size_t MatchingAddressCount() const noexcept;
 
-    struct Result
-    {
-        ra::ByteAddress nAddress{};
-        unsigned int nValue{};
-        MemSize nSize{};
-    };
-
     /// <summary>
     /// Initializes a result set from a list of address/value pairs
     /// </summary>
     /// <param name="srResults">The results to populate from.</param>
     /// <param name="nType">Type of search to initialize.</param>
-    void Initialize(_In_ const std::vector<Result>& srResults, _In_ SearchType nSearchType);
+    void Initialize(_In_ const std::vector<SearchResult>& srResults, _In_ SearchType nSearchType);
 
     /// <summary>
     /// Gets the nIndex'th matching address.
     /// </summary>
     /// <param name="result">The result.</param>
     /// <returns><c>true</c> if result was populated, <c>false</c> if the index was invalid.</returns>
-    bool GetMatchingAddress(gsl::index nIndex, _Out_ Result& result) const noexcept;
+    bool GetMatchingAddress(gsl::index nIndex, _Out_ SearchResult& result) const noexcept;
 
     /// <summary>
     /// Gets and item from the results matching a result from another set of results.
@@ -268,7 +124,7 @@ public:
     /// <param name="pSrcResult">The result to find a match for.</param>
     /// <param name="result">The result.</param>
     /// <returns><c>true</c> if result was populated, <c>false</c> if the index was invalid.</returns>
-    bool GetMatchingAddress(const SearchResults::Result& pSrcResult, _Out_ Result& result) const noexcept;
+    bool GetMatchingAddress(const SearchResult& pSrcResult, _Out_ SearchResult& result) const noexcept;
 
     /// <summary>
     /// Gets the raw bytes at the specified address.
@@ -290,7 +146,7 @@ public:
     /// <param name="pPreviousResult">The previous value of the result.</param>
     /// <param name="sFormattedValue">Pointer to a string to populate with a textual representation of the value. [optional]</param>
     /// <returns><c>true</c> if the value changed, <c>false</c> if not.</returns>
-    bool UpdateValue(SearchResults::Result& pResult, _Out_ std::wstring* sFormattedValue,
+    bool UpdateValue(SearchResult& pResult, _Out_ std::wstring* sFormattedValue,
         const ra::data::context::EmulatorContext& pEmulatorContext) const;
 
     /// <summary>
@@ -299,7 +155,7 @@ public:
     /// <param name="pPreviousResults">The previous results to compare against.</param>
     /// <param name="pResult">The result to test.</param>
     /// <returns><c>true</c> if the value still matches the filter, <c>false</c> if not.</returns>
-    bool MatchesFilter(const SearchResults& pPreviousResults, SearchResults::Result& pResult) const;
+    bool MatchesFilter(const SearchResults& pPreviousResults, SearchResult& pResult) const;
 
     /// <summary>
     /// Gets the type of search performed.
@@ -345,16 +201,16 @@ public:
     /// <param name="pResult">The result to remove.</param>
     /// <returns><c>true</c> if the specified address was removed from the matching address list;
     /// otherwise <c>false</c>.</returns>
-    bool ExcludeResult(const SearchResults::Result& pResult);
+    bool ExcludeResult(const SearchResult& pResult);
 
 private:
     void MergeSearchResults(const SearchResults& srMemory, const SearchResults& srAddresses);
 
-    std::vector<impl::MemBlock> m_vBlocks;
+    std::vector<search::MemBlock> m_vBlocks;
     SearchType m_nType = SearchType::EightBit;
 
-    friend class impl::SearchImpl;
-    impl::SearchImpl* m_pImpl = nullptr;
+    friend class search::SearchImpl;
+    search::SearchImpl* m_pImpl = nullptr;
 
     ComparisonType m_nCompareType = ComparisonType::Equals;
     SearchFilterType m_nFilterType = SearchFilterType::None;
