+++ tests/RA_Integration.Tests.vcxproj.filters
@@ -486,6 +486,12 @@
     <ClCompile Include="services\AchievementLogicSerializer_Tests.cpp">
       <Filter>Tests\Services</Filter>
     </ClCompile>
+    <ClCompile Include="..\src\services\search\MemBlock.cpp">
+      <Filter>Code</Filter>
+    </ClCompile>
+    <ClCompile Include="..\src\services\search\SearchImpl.cpp">
+      <Filter>Code</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <None Include="base.props" />
